requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
pydantic>=2.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0
structlog>=23.0.0
wandb>=0.15.0
pyyaml>=6.0
